requirements.txt
groq
gradio
reportlab
markdown2
python-dotenv
numpy
faiss-cpu
ollama